Create Cloud Provider Equinix Metal add-on
The new add-on installs and configures Cloud Provider Equinix Metal
on the target kubernetes cluster.  For now, the installation is via
SSH.

Installing via SSH assumes that there is a remote node that has the
proper access & configuration for running `kubectl` commands. It might
be nicer to instead expect customers to do whatever setup is needed in
order to use the `kubectl` or `kubernetes` terraform provider so that
we can get stateful tracking of cluster changes.

The CPEM add-on exposes device provisioning prerequisites that can
be included in the userdata for the machines in a kubernetes cluster
to ensure that the machines/cluster are configured properly for CPEM.
Prerequisites are a proof-of-concept and may be removed in the future.