fix: remove requirement for unused template provider
The portworx and portworx-data-services modules declare a requirement
for the (deprecated) template provider, but they do not include any
resources from that provider.

The provider requirement makes it impossible to init a terraform config
that uses this module on an ARM-based Mac, because the template provider
was deprecated before that platform existed, so there is no compatible
build of that provider.